requirements.txt
# INSERT HERE THE PACKAGES NEEDED

fastapi==0.110.2
imblearn==0.0
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
pyyaml==6.0.1
scikit-learn==1.3.0
scipy==1.13.0
seaborn==0.11.1
shap==0.41.0
starlette==0.37.2
statsmodels==0.14
uvicorn==0.29.0